Modified split_args* so it saves the quotation marks of strings
The previous version of the function deleted the quotation marks in strings,
so examples like [">>outfile"] which is a string would be converted to
[>>outfile] and be interpreted as a redirection. Also when using the $
to specify an argument, it was not posible to know whether it was inside single
quotes and ignore it, or in double quotes or outside them to evaluate the
variable.
These problems have been fixed by saving the quotation marks in the strings.