Fixed segv and leaks from main.
When an input with variables that expanded into an empty string was written in
the minishell, there was a segmentation fault because the parse returns a list
with the first command being empty. Also if double quotes are written, the parse
returns that the command first argument is empty, and that gave problems too.