Updated src/get_variable.c to diferentiate certain cases.
The case where the variable content is NULL and there is no variable with
a certain name was the same. The function has been changed so it returns 1 if
it finds the name, and 0 if not. The content is returned to a new parameter of
the function that is a reference to a string.