Modified get_variable so it returns a strdup'd string.
This is a breaking change, but it's a good one.  It's a breaking change
because the caller is now responsible for freeing the string returned by
get_variable.  This is a good thing because it means that the caller can
now use the string after the next call to get_variable.
Also, the return values of get_variable have been updated. If the name is
not found in the list, 0 is returned. If it is found and the strdup was
succesful, 1 is returned. If it is found and the strdup failed, -1 is
returned.