Updated exit builtin so it works like the original
Added the option to pass 1 argument to exit.
If no argument is passed the shell will exit with the value of the last
	command executed.
If one argument is passed the shell will exit with the value of the argument.
If this argument is not a number "[+-][0-9]+", the shell will print an error
	message and exit with status code 2.
If 2 arguments are passed, the shell will print an error message and continue
	running, setting $? to 1.