PARSE: Finished expand_variables function and some minor changes
The new expand_variables function is now working as expected. It takes a list
containing t_arg_redir (2 lists of arguments), and expands the variables
contained in the arguments. It also checks for the error of ambiguous redirect
and prints the error message if needed.
Also free_arg_redir has been modified so it doesn't crash if it receives a NULL